+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+## 0.2.0 - 2024-04-03
+### Added
+- Added `vm/vars` DELETE Method to clear all user-defined variables
+- Added `vm/var` DELETE Method to clear a single user-defined variable
+
 ## 0.1.0 - 2024-04-02
 ### Added
 - Added `vm/vars` POST Method to set all user-defined variables